refactor flint package
